+++ fishpondserver/src/main/resources/application.properties
@@ -2,7 +2,7 @@
 server.port=8080
 
 #	数据库url	
-spring.datasource.url=jdbc:postgresql://124.70.67.6:26000/fishpond?ApplicationName=app1
+spring.datasource.url=jdbc:postgresql://124.70.67.6:26000/fishpond?useAffectedRows=true
 #	数据库用户名	
 spring.datasource.username=mhn
 #	数据库密码
