+++ fishpondserver/src/test/java/edu/bit/fishpond/DAOTest.java
@@ -1,7 +1,7 @@
 package edu.bit.fishpond;
 
-import edu.bit.fishpond.mapper.HumanMapper;
-import edu.bit.fishpond.mapper.UserInfoMapper;
+import edu.bit.fishpond.DAO.mapper.HumanMapper;
+import edu.bit.fishpond.DAO.mapper.UserInfoMapper;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -21,6 +21,7 @@ public class DAOTest extends FishpondApplicationTests {
     @Test
     public void testUserInfo() {
 
+        System.out.println(userInfoMapper.resetSeq());
         System.out.println(userInfoMapper.insertOne("刘睿",
                 "M",
                 "2000-01-01",
@@ -30,9 +31,9 @@ public class DAOTest extends FishpondApplicationTests {
                 "2021-08-30",
                 false));
         System.out.println(userInfoMapper.selectAll().toString());
-        System.out.println(userInfoMapper.updateOne("region","湖北",1));
-        System.out.println(userInfoMapper.updateOne("state","true",1));
-        System.out.println(userInfoMapper.selectOne(1).toString());
+        //System.out.println(userInfoMapper.updateOne("region","湖北",1));
+        //System.out.println(userInfoMapper.updateOne("state","true",1));
+        //System.out.println(userInfoMapper.selectOne(1).toString());
         System.out.println(userInfoMapper.deleteAll());
     }
 }
